Sweep the last five bugprone checks and make the tier binding (#2530)
* Sweep the last five bugprone checks and make the tier binding

The five checks the pre-push gate still excluded are at zero, so
WarningsAsErrors carries 'bugprone-*,cert-*' from here.

Seven of the 19 widening findings were literal products that fit in int and
fold at compile time (64 * 1024, 2 * 1024 * 1024, 24 * 60 * 60), so they are
config rather than code: IgnoreConstantIntExpr. The other twelve are runtime
int products widened to an index, and take a cast. unused-return-value is
AllowCastToVoid, since static_cast<void>(p.release()) is how both sites
already said the discard was deliberate.

The reserved-identifier and return-const-ref NOLINTs were both dead. Each sat
above a comment that wrapped onto a second line, so it suppressed the comment
rather than the code, and return-const-ref reports on the return statement
rather than on the declaration the marker sat above. The four deliberate
ClipWaveformPainter roundings now go through one named helper carrying a
single suppression, instead of four.

The five checks #2529 added to the gate's exclusion list are removed: no
check by any of those names exists. The real spellings are
hicpp-signed-bitwise, cert-flp30-c and cert-err34-c, none of them reachable
from a bugprone-* allow-list, so the lines excluded nothing and the sample
counts recorded against them measured something else.


* Free the playback context before the test engine closes its device

getSharedEngine() closes the audio device so nothing in a test binary reaches
the speakers, but initialize() has just called ensureContextAllocated().
closeDevices() asserts activeContexts.isEmpty(), so every run fired a jassert
at init.

The corpus harness reads an assertion during a case as "the engine asserted
while rendering" and never forgives one, so this surfaced as two unrelated
looking case failures -- placement.grid and project.v1bounces -- against
whichever case happened to trigger the lazy init. Both corpora are now clean:
5/5 and 9/9.

Shutdown already frees the context before closing devices; this is the same
order.


* Keep the audio-device warning asynchronous

showMessageBoxAsync takes JUCE's runSync() path when the callback is null, so
this "async" warning ran a modal loop on the message thread. play() is not
only a user action -- tests, controllers and the MCP transport_play API all
call it -- so the loop could open with nobody able to press OK.

It hung the single-process JUCE run indefinitely, and only there: the branch
needs an earlier suite to have left a device playing, so running the suites
one process each passed. An empty callback is what selects runAsync.


* Download sccache to a file, and retry it

The Linux build failed on this branch having compiled nothing: curl's
connection was reset partway through the sccache tarball, gzip hit an
unexpected end of file and tar exited 2.

Piping curl into tar cannot survive that. --retry resends the request, but the
truncated bytes have already gone down the pipe, so tar is reading a corrupt
archive either way. Downloading to a file first makes a retry mean something.
All three sccache fetches get the same treatment, Windows included, where the
download was already to a file but had no retry.

+++ magda/agents/command_model.cpp
@@ -6,6 +6,7 @@
 #include <array>
 #include <cctype>
 #include <cmath>
+#include <cstddef>
 #include <cstdio>
 #include <cstdlib>
 #include <stdexcept>
@@ -650,16 +651,16 @@ CommandModel::Prediction CommandModel::predict(const std::string& text) const {
     // Embedding → [E][ML] (channel-major, matching the .transpose(1,2)).
     std::vector<float> e(E * ML, 0.0f);
     for (int t = 0; t < ML; ++t) {
-        const float* row = &d::kEmbedWeight[ids[t] * E];
+        const float* row = &d::kEmbedWeight[static_cast<std::size_t>(ids[t]) * E];
         for (int c = 0; c < E; ++c)
             e[c * ML + t] = row[c];
     }
 
     auto conv = [&](const std::vector<float>& in, int ci, int co, const float* W, const float* B,
                     int dilation) {
-        std::vector<float> out(co * ML, 0.0f);
+        std::vector<float> out(static_cast<std::size_t>(co) * ML, 0.0f);
         for (int oc = 0; oc < co; ++oc) {
-            const float* wOc = &W[oc * ci * 3];
+            const float* wOc = &W[static_cast<std::size_t>(oc) * ci * 3];
             for (int t = 0; t < ML; ++t) {
                 float acc = B[oc];
                 for (int k = 0; k < 3; ++k) {
@@ -686,7 +687,7 @@ CommandModel::Prediction CommandModel::predict(const std::string& text) const {
         float bestVal = 0.0f;
         for (int tag = 0; tag < d::kNumTags; ++tag) {
             float acc = d::kSlotBias[tag];
-            const float* w = &d::kSlotWeight[tag * H];
+            const float* w = &d::kSlotWeight[static_cast<std::size_t>(tag) * H];
             for (int c = 0; c < H; ++c)
                 acc += w[c] * h[c * ML + t];
             if (tag == 0 || acc > bestVal) {
@@ -709,7 +710,7 @@ CommandModel::Prediction CommandModel::predict(const std::string& text) const {
     float bestIntentVal = 0.0f;
     for (int i = 0; i < d::kNumIntents; ++i) {
         float acc = d::kIntentBias[i];
-        const float* w = &d::kIntentWeight[i * H];
+        const float* w = &d::kIntentWeight[static_cast<std::size_t>(i) * H];
         for (int c = 0; c < H; ++c)
             acc += w[c] * pooled[c];
         if (i == 0 || acc > bestIntentVal) {
